Begin creating auth functionality
Created  basic HTML for sign up, sign in, change PW, and sign out.
Added event listeners for all four
Added AJAX calls for all 4.
Created success and fail functions for sign up, sign in, and sign out.
On successful sign up, credentials are passed to sign in function to
allow for auto-sign in.