Create update item functionality
Add event listeners for enter key and for focusout on list items.
By grabbing the id of the item, the previous values, and which part
of the item(name or quantity) was targeted, I pass several arguments
to the api and update the item. Reload 'get' on success to dynamically
update.