Fix bug in edit items
Because another element could be focused on before submitting edits
to the current item, it was causing wonky things to happen.
After much testing, the solution was to include a class 'edit'
on each editable field. Then on focus, remove that class from the
event target, but remove the ability to edit all items with this class.
This works because on completio of this function, the items are newly
generated with the ability to be edited again